Separate router
Move router out of index
